requirements.txt
rpaframework
pydantic-settings
confluent-kafka
apache-airflow-providers-apache-kafka